Refactor: move compare-features script test out of linter (#9229)
* Refactor: move compare-features test out of linter (bad indentation)

* Run npm run fix

* Feedback
