+++ test/lint.js
@@ -16,7 +16,6 @@ const {
   testDescriptions,
 } = require('./linter/index.js');
 const { IS_CI } = require('./utils.js');
-const testCompareFeatures = require('./test-compare-features');
 const testMigrations = require('./test-migrations');
 const testFormat = require('./test-format');
 
@@ -156,7 +155,6 @@ var hasErrors = argv.files
       'webdriver',
       'webextensions',
     );
-hasErrors = testCompareFeatures() || hasErrors;
 hasErrors = testMigrations() || hasErrors;
 hasErrors = testFormat() || hasErrors;
